src/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "Doc-Search"
version = "0.0.1"
description = "Local free text search"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE.txt"}
keywords = ["text-search", "embeddings", "offline"]
authors = [
  {name = "Lukas Materne", email = "lmaterne@proton.me" }
]
maintainers = [
  {name = "Lukas Materne", email = "lmaterne@proton.me" }
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Anyone",
  "Topic :: Academic Tools :: Search Tools",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
  "fastembed",
  "python-docx",
  "PyPDF2",
]

[project.optional-dependencies]
dev = ["check-manifest"]
test = ["coverage"]

[project.urls]
"Homepage" = "https://github.com/LMaterne/doc-search"
"Bug Reports" = "https://github.com/LMaterne/doc-searchissues"
"Source" = "https://github.com/LMaterne/doc-search"

[tool.setuptools]
py-modules = []
